+++ src/main/java/com/trustamarket/paymentservice/paymentservice/application/dto/command/PaymentDetailCommand.java
@@ -1,0 +1,8 @@
+package com.trustamarket.paymentservice.paymentservice.application.dto.command;
+
+import java.util.UUID;
+
+public record PaymentDetailCommand(
+        UUID paymentId,
+        UUID userId
+) {}
