+++ src/main/java/com/trustamarket/paymentservice/paymentservice/domain/repository/condition/PaymentSearchCondition.java
@@ -1,10 +1,10 @@
-package com.trustamarket.paymentservice.paymentservice.application.dto.query;
+package com.trustamarket.paymentservice.paymentservice.domain.repository.condition;
 
 import com.trustamarket.paymentservice.paymentservice.domain.enums.PaymentStatus;
 
 import java.util.UUID;
 
-public record SearchPaymentQuery(
+public record PaymentSearchCondition(
         UUID paymentId,
         UUID userId,
         long minAmount,
